refactor: permissions package structure in iosMain
