+++ identity_presentation/src/iosMain/kotlin/net/thechance/mena/identity/presentation/di/platformModule.ios.kt
@@ -1,9 +1,9 @@
 package net.thechance.mena.identity.presentation.di
 
 import net.thechance.mena.identity.presentation.util.AppLocalizer
-import net.thechance.mena.identity.presentation.util.CameraPermission
-import net.thechance.mena.identity.presentation.util.GalleryPermission
-import net.thechance.mena.identity.presentation.util.LocationForegroundPermission
+import net.thechance.mena.identity.presentation.util.permissions.CameraPermission
+import net.thechance.mena.identity.presentation.util.permissions.GalleryPermission
+import net.thechance.mena.identity.presentation.util.permissions.LocationForegroundPermission
 import net.thechance.mena.identity.presentation.util.permissionHandler.PermissionController
 import net.thechance.mena.identity.presentation.util.permissionHandler.Permissions
 import org.koin.core.module.Module
