--- identity_presentation/src/iosMain/kotlin/net/thechance/mena/identity/presentation/util/CameraPermission.kt
@@ -1,58 +1,0 @@
-package net.thechance.mena.identity.presentation.util
-
-import kotlinx.coroutines.suspendCancellableCoroutine
-import net.thechance.mena.identity.domain.exception.PermissionDeniedException
-import net.thechance.mena.identity.domain.exception.PermissionDeniedPermanentlyException
-import net.thechance.mena.identity.domain.exception.PermissionNotDeterminedException
-import net.thechance.mena.identity.presentation.util.permissionHandler.PermissionController
-import net.thechance.mena.identity.presentation.util.permissionHandler.PermissionState
-import platform.AVFoundation.AVAuthorizationStatus
-import platform.AVFoundation.AVAuthorizationStatusAuthorized
-import platform.AVFoundation.AVAuthorizationStatusDenied
-import platform.AVFoundation.AVAuthorizationStatusNotDetermined
-import platform.AVFoundation.AVAuthorizationStatusRestricted
-import platform.AVFoundation.AVCaptureDevice
-import platform.AVFoundation.AVMediaTypeVideo
-import platform.AVFoundation.authorizationStatusForMediaType
-import platform.AVFoundation.requestAccessForMediaType
-import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
-
-internal class CameraPermission: PermissionController {
-    override fun getPermissionState(): PermissionState {
-        val status = AVCaptureDevice.authorizationStatusForMediaType(AVMediaTypeVideo)
-        return status.toPermissionState()
-    }
-
-    override fun openSettingPage() {
-        openAppSettingsPage()
-    }
-
-    override suspend fun requestPermission() {
-        return suspendCancellableCoroutine { cont ->
-            AVCaptureDevice.requestAccessForMediaType(AVMediaTypeVideo) { granted ->
-                if (cont.isCancelled) return@requestAccessForMediaType
-                if (granted) {
-                    cont.resume(Unit)
-                } else {
-                    when (getPermissionState()) {
-                        PermissionState.GRANTED -> cont.resume(Unit)
-                        PermissionState.NOT_DETERMINED -> cont.resumeWithException(PermissionNotDeterminedException())
-                        PermissionState.DENIED -> cont.resumeWithException(PermissionDeniedException())
-                        PermissionState.DENIED_PERMANENTLY -> cont.resumeWithException(PermissionDeniedPermanentlyException())
-                    }
-                }
-            }
-        }
-    }
-
-    private fun AVAuthorizationStatus.toPermissionState(): PermissionState {
-        return when (this) {
-            AVAuthorizationStatusAuthorized -> PermissionState.GRANTED
-            AVAuthorizationStatusNotDetermined -> PermissionState.NOT_DETERMINED
-            AVAuthorizationStatusRestricted -> PermissionState.DENIED_PERMANENTLY
-            AVAuthorizationStatusDenied -> PermissionState.DENIED
-            else -> PermissionState.NOT_DETERMINED
-        }
-    }
-}
